[SPARK-55802][SQL] Fix integer overflow when computing Arrow batch bytes
### What changes were proposed in this pull request?

`ArrowWriter.sizeInBytes()` and `SliceBytesArrowOutputProcessorImpl .getBatchBytes()` both accumulated per-column buffer sizes (each an `Int`) into an `Int` accumulator. When the total exceeds 2 GB the sum silently wraps negative, causing the byte-limit checks controlled by `spark.sql.execution.arrow.maxBytesPerBatch` and
`spark.sql.execution.arrow.maxBytesPerOutputBatch` to behave incorrectly and potentially allow oversized batches through.

Fix by changing both accumulators and return types to `Long`.

### Why are the changes needed?

Fix possible overflow when calculating Arrow batch bytes.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude Sonnet 4.6 <noreplyanthropic.com>

Closes #54584 from viirya/fix-arrow-batch-bytes-overflow.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
(cherry picked from commit df195ac59de5bd896cd70699cfe96ebf78bf2976)

+++ sql/core/src/test/scala/org/apache/spark/sql/execution/python/streaming/BaseStreamingArrowWriterSuite.scala
@@ -76,7 +76,7 @@ class BaseStreamingArrowWriterSuite extends SparkFunSuite with BeforeAndAfterEac
       ()
     }
 
-    when(arrowWriter.sizeInBytes()).thenAnswer { _ => sizeCounter }
+    when(arrowWriter.sizeInBytes()).thenAnswer { _ => sizeCounter.toLong }
 
     // Set arrowMaxBytesPerBatch to 1
     transformWithStateInPySparkWriter = new BaseStreamingArrowWriter(
